optimal-route-kl/backend/requirements-flask.txt
Flask==3.0.0
flask-cors==4.0.0
googlemaps==4.6.0
python-dotenv==1.0.0
requests==2.32.5
Werkzeug==3.0.0

